+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Inspired from [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 - expose chatEnabled flag to capabilities ([#398](https://github.com/opensearch-project/dashboards-assistant/pull/398))
 - update chatbot UI to align with new look ([#435](https://github.com/opensearch-project/dashboards-assistant/pull/435))
 - add flag to control if display conversation list ([#438](https://github.com/opensearch-project/dashboards-assistant/pull/438))
+- add index type detection ([#454](https://github.com/opensearch-project/dashboards-assistant/pull/454))
 
 ### Enhancements